Kill foreground process by long-pressing BACK
This patch allows the user to force the running foreground application to be
killed. One example use case for this is applications which trap buttons,
leaving the user stuck. Another would be a memory-heavy application that
traps the BACK button (so they user wouldn't want to leave it in memory by
pressing HOME). Browser is a good example of this second use case.

Change-Id: Iaa7fb78fbb6129ae01066651e97eae5f0ef527b6